backend/requirements.txt
# ============================================
# GB Guide — Backend Dependencies
# ============================================

# Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
sqlmodel==0.0.22

# Migrations
alembic==1.14.1

# Environment
python-dotenv==1.0.1

# Utilities
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# Authentication
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
